+++ src/utils/point.js
@@ -1,13 +1,5 @@
 import dayjs from 'dayjs';
 
-function getRandomArrayElement(items) {
-  return items[Math.floor(Math.random() * items.length)];
-}
-
-function getRandomInteger(integer) {
-  return Math.floor(Math.random() * integer);
-}
-
 function humanizeTaskDueDate(dueDate, dateFormat) {
   return dueDate ? dayjs(dueDate).format(dateFormat) : '';
 }
@@ -16,9 +8,6 @@ function calculatesTravelTime(dateFrom, dateTo) {
   const date1 = dayjs(dateTo);
   return date1.diff(dateFrom, 'minute');
 }
-
-const isEscapeKey = (evt) => evt.keyCode === 27;
-
 
 function createFormOffersTemplate(pointOffers, point) {
   return pointOffers
@@ -46,4 +35,4 @@ function createEventTypeItem (offers) {
 </div>`).join('');
 }
 
-export {getRandomArrayElement, getRandomInteger, humanizeTaskDueDate, calculatesTravelTime, isEscapeKey, createFormOffersTemplate, createDestinationList, createEventTypeItem};
+export {humanizeTaskDueDate, calculatesTravelTime, createFormOffersTemplate, createDestinationList, createEventTypeItem};
