+++ src/view/creation-form-view.js
@@ -1,5 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import {humanizeTaskDueDate, createFormOffersTemplate, createDestinationList, createEventTypeItem} from '../util.js';
+import {humanizeTaskDueDate, createFormOffersTemplate, createDestinationList, createEventTypeItem} from '../utils/point.js';
 import {DATE_TIME_FORMAT} from '../const.js';
